DAT2000_V24_oblig-main/requirements.txt
polars==0.20.14
pandas==2.2.1
uvicorn==0.28.0
fastapi==0.110.0
sqlalchemy==2.0.28
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
pyarrow==15.0.1